+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 1.3.3 BAngles calculation
+- BAngles can now be calculated without having to perform the remap
+- netCDF4 files can be read and analysed
+- The time instant can be calculated for the absolute or relative frame number
+- MU default position updated
+- Camera general parameters can now be read
+
+
 # 1.3.2 Minor polishing
 - `timeWindow` option of the get_ECE function can now be a tuple
 - Add recursive finding to read Calibration database to alow having more header lines than the Suite standard